Split queue and stream into separate benchmarks
